separate out autocomplete component
