+++ context/app/static/js/components/workspaces/WorkspacesList.tsx
@@ -1,6 +1,5 @@
 import React, { useMemo, useState } from 'react';
 
-import { Autocomplete, AutocompleteRenderInputParams, Box, InputAdornment } from '@mui/material';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 import DeleteRounded from '@mui/icons-material/DeleteRounded';
@@ -10,8 +9,7 @@ import { useSelectItems } from 'js/hooks/useSelectItems';
 import WorkspaceButton from 'js/components/workspaces/WorkspaceButton';
 import NewWorkspaceDialogFromWorkspaceList from 'js/components/workspaces/NewWorkspaceDialog/NewWorkspaceDialogFromWorkspaceList';
 import WorkspacesTable from 'js/components/workspaces/Tables/WorkspacesTable';
-import { WorkspaceWithUserId } from 'js/components/workspaces/types';
-import { StyledSearchIcon, StyledTextField } from 'js/components/workspaces/style';
+import WorkspacesAutocomplete from 'js/components/workspaces/WorkspacesAutocomplete';
 import { useWorkspacesListWithSharerInfo } from './hooks';
 import ConfirmDeleteWorkspacesDialog from './ConfirmDeleteWorkspacesDialog';
 
@@ -43,69 +41,20 @@ function ShareWorkspaceButton({ selectedItems }: { selectedItems: Set<string> })
   );
 }
 
-function WorkspaceOption(props: React.HTMLAttributes<HTMLLIElement>, option: WorkspaceWithUserId) {
-  const { name, id } = option;
-
-  return (
-    <li {...props} key={id}>
-      <div>
-        <Typography variant="subtitle1">{name}</Typography>
-        <Typography variant="body2">{`ID: ${id}`}</Typography>
-      </div>
-    </li>
-  );
-}
-
-function WorkspaceSearchField(params: AutocompleteRenderInputParams) {
-  const { InputProps } = params;
-  return (
-    <StyledTextField
-      {...params}
-      fullWidth
-      size="small"
-      placeholder="Search workspace by name or ID"
-      slotProps={{
-        input: {
-          ...InputProps,
-          startAdornment: (
-            <InputAdornment position="start">
-              <StyledSearchIcon />
-            </InputAdornment>
-          ),
-        },
-      }}
-    />
-  );
-}
-
-function useWorkspacesAutocomplete({ workspaces = [] }: { workspaces: WorkspaceWithUserId[] }) {
-  const [inputValue, setInputValue] = useState('');
-  const autocompleteValue = '';
-
-  const filteredWorkspaces = useMemo(() => {
-    if (!inputValue.trim()) return workspaces;
-    return workspaces.filter(
-      (ws) =>
-        ws.name.toLowerCase().includes(inputValue.toLowerCase()) || ws.id.toString().includes(inputValue.toLowerCase()),
-    );
-  }, [inputValue, workspaces]);
-
-  return {
-    inputValue,
-    setInputValue,
-    autocompleteValue,
-    filteredWorkspaces,
-  };
-}
-
 function WorkspacesList() {
   const { workspacesList, handleDeleteWorkspace, isDeleting, isLoading } = useWorkspacesListWithSharerInfo();
   const { selectedItems, toggleItem } = useSelectItems();
   const [dialogIsOpen, setDialogIsOpen] = useState(false);
 
-  const { inputValue, setInputValue, autocompleteValue, filteredWorkspaces } = useWorkspacesAutocomplete({
-    workspaces: workspacesList,
-  });
+  const [inputValue, setInputValue] = useState('');
+
+  const filteredWorkspaces = useMemo(() => {
+    if (!inputValue.trim()) return workspacesList;
+    return workspacesList.filter(
+      (ws) =>
+        ws.name.toLowerCase().includes(inputValue.toLowerCase()) || ws.id.toString().includes(inputValue.toLowerCase()),
+    );
+  }, [inputValue, workspacesList]);
 
   return (
     <>
@@ -119,22 +68,11 @@ function WorkspacesList() {
       <Stack spacing={2}>
         <Typography variant="h4">{`My Workspaces (${workspacesList.length})`}</Typography>
         <Stack spacing={1} direction="row" alignItems="center" justifyContent="space-between">
-          <Box flex={1} maxWidth="50%">
-            <Autocomplete
-              value={autocompleteValue}
-              inputValue={inputValue}
-              onInputChange={(event, newInputValue) => {
-                setInputValue(newInputValue);
-              }}
-              filterOptions={(x) => x}
-              options={filteredWorkspaces}
-              renderOption={WorkspaceOption}
-              getOptionLabel={(option) => (typeof option === 'string' ? option : option.name)}
-              freeSolo
-              fullWidth
-              renderInput={WorkspaceSearchField}
-            />
-          </Box>
+          <WorkspacesAutocomplete
+            inputValue={inputValue}
+            setInputValue={setInputValue}
+            filteredWorkspaces={filteredWorkspaces}
+          />
           <Stack display="flex" direction="row" spacing={2}>
             <DeleteWorkspaceButton
               disabled={isDeleting || selectedItems.size === 0}
